feat(collector-tag-form): create component
