+++ apps/web/src/services/asset-inventory/collector/modules/CollectorTagForm.vue
@@ -1,0 +1,60 @@
+<template>
+    <!--            TODO: translation-->
+    <p-field-group class="collector-tag-form"
+                   :label="$t('Tag')"
+    >
+        <template #label-extra>
+            <div class="mt-1">
+                <!-- TODO: translation -->
+                <p class="tag-description">
+                    {{ $t("Set Account's tag.") }}
+                </p>
+                <p class="tag-description">
+                    {{ $t("The Key - Value pair is a required field. Only underscores (_), characters, and numbers are allowed. International characters are allowed.") }}
+                </p>
+            </div>
+        </template>
+        <tags-input-group :tags="state.tags"
+                          show-validation
+                          :is-valid.sync="state.isTagsValid"
+                          @update-tags="handleUpdateTags"
+        />
+    </p-field-group>
+</template>
+
+<script lang="ts" setup>
+
+import { reactive, watch } from 'vue';
+
+import { PFieldGroup } from '@spaceone/design-system';
+
+import TagsInputGroup from '@/common/components/forms/tags-input-group/TagsInputGroup.vue';
+import type { Tag } from '@/common/components/forms/tags-input-group/type';
+
+const emit = defineEmits<{(isTagsValid: 'updateIsTagsValid', value: boolean): void; }>();
+
+const state = reactive({
+    tags: {},
+    isTagsValid: true,
+});
+
+/* event */
+const handleUpdateTags = (tags: Tag) => {
+    state.tags = tags;
+};
+
+watch(() => state.isTagsValid, (isTagsValid) => {
+    emit('updateIsTagsValid', isTagsValid);
+});
+
+</script>
+
+<style lang="postcss" scoped>
+.collector-tag-form {
+    .tag-description {
+        @apply text-label-md text-gray-900;
+        font-weight: 400;
+    }
+}
+
+</style>
